Added Address Repo and created service logic for POST student Address

+++ src/main/java/com/project/StudentManagement/dto/AddressDTO.java
@@ -1,5 +1,6 @@
 package com.project.StudentManagement.dto;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,5 +14,5 @@ public class AddressDTO {
     private Integer id;
     private String locality;
     private String city;
-    private Long pincode;
+    private Integer pincode;
 }
